+++ community/c++-algo/include/GameLib/src/gameMap.h
@@ -31,8 +31,8 @@ namespace terminal {
         bool inArenaBounds(Pos pos) const;
         void getEdgeLocations(vector<Pos>& vec, const EDGE edge) const;
         void getEdges(vector<vector<Pos> >& vec);
-        void addUnit(UNIT_TYPE unitType, int x, int y, int playerIndex, int hp = 0);
-        void addUnit(UNIT_TYPE unitType, Pos pos, int playerIndex, int hp = 0);
+        void addUnit(UNIT_TYPE unitType, int x, int y, int playerIndex, double hp = 0);
+        void addUnit(UNIT_TYPE unitType, Pos pos, int playerIndex, double hp = 0);
         void removeUnits(Pos pos);
         void getLocationsInRange(vector<Pos>& locations, Pos pos, double radius);
         bool containsStationaryUnit(Pos pos) const;
